packages/partner-api/requirements.txt
fastapi
uvicorn
pymongo
pydantic-settings
python-jose
sqlalchemy
